requirements.txt
# Delta: Self-Extensible Agent System

# Core LLM
google-generativeai>=0.8.0

# Sandboxed Execution
RestrictedPython>=7.0

# Data Validation
pydantic>=2.0

# Async Storage
aiosqlite>=0.19.0

# HTTP Client
httpx>=0.27.0
